fst1: make letters to numbers
This FST can do the following.

Step 1: Retain the first letter of the name. This may be uppercased or
lowercased.

Step 2: Remove all non-initial occurrences of the following letters: a,
e, h, i, o, u, w, y. (To clarify, this step removes all occurrences of
the given characters except when they occur in the first position.)

Step 3: Replace the remaining letters (except the first) with numbers:

• b, f, p, v → 1
• c, g, j, k, q, s, x, z → 2 • d, t → 3
•l→4
• m, n → 5
•r→6

If two or more letters from the same number group were adjacent in the
original name, then only replace the first of those letters with the
corresponding number and ignore the others.